% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_diversity_plot.R
\name{basic.div}
\alias{basic.div}
\title{This function returns a plot to show the composition and diversity of samples}
\usage{
basic.div(data, threshold_rare = 0.03)
}
\arguments{
\item{data}{a data table with each row corresponding to a pair of category and
sample. The table must also have columns for relative and/or absolute abundances}

\item{threshold_rare}{some categories might contain very few organisms which
can be grouped together in a new category : 'other'. By default this threshold
is set to 0.03 (3\%), which corresponds to the proportion of organisms represented
by this category. If you do not want to group the least abundant categories, set
this threshold to 0.}
}
\value{
an object of type ggplot
}
\description{
This function returns a plot to show the composition and diversity of samples
}
