% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovolumes.R
\name{biovolume}
\alias{biovolume}
\title{This function calculates the biovolume of a particle. This is the percentage
volume a particle \link{mm^3} takes up in a unit of filtered water\link{m^3}.}
\usage{
biovolume(samples, objects)
}
\arguments{
\item{samples}{A data table containing samples in rows and details about
the latter in columns.}

\item{objects}{A data table containing objects in rows (not necessarily
from the same sample) and details about the objects in the columns.}
}
\description{
This function calculates the biovolume of a particle. This is the percentage
volume a particle \link{mm^3} takes up in a unit of filtered water\link{m^3}.
}
\note{
This function does not return anything, it adds directly a column
containing the biovolume of each object in the data table objects

It is relevant to talk about biovolume when looking about at a certain
catgeory of plankton, when we sum individual biovolumes to show what part
certain species take up in a unit of water.
}
