% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankto_data_utils.R
\name{get.counts.and.vol}
\alias{get.counts.and.vol}
\title{Extract total counts and volumes}
\usage{
get.counts.and.vol(planktotable)
}
\arguments{
\item{planktotable}{a data table obtained using the function import.data()}
}
\value{
A data table with 5 columns. The sample id, object category, counts,
volume and biovolumes.
}
\description{
This function counts how many objects has been observed
per category (object_annotation_category) for each sample. It also
sums the volumes and biovolumes for each category.
}
