% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankto_data_utils.R
\name{get.df.vegan}
\alias{get.df.vegan}
\title{Get a community data matrix}
\usage{
get.df.vegan(x, value.var = "count")
}
\arguments{
\item{x}{a long format table obtained using the function get.counts.and.vol()}

\item{value.var}{value column to used. Possible options are "count", "volume"
and "biovolume"}
}
\value{
A community data matrix (data frame) with categories as columns and samples as rows.
}
\description{
This function convert a long format table to a community
data matrix that can be directly used by functions from the library vegan
}
