% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankto_data_utils.R
\name{get.info}
\alias{get.info}
\title{Get description of samples or objects}
\usage{
get.info(planktotable, type = "sample")
}
\arguments{
\item{planktotable}{a data table obtained using the function import.data()}

\item{type}{type of fields to extract. Possible options are "sample" and "object"}
}
\value{
A data tale with information related to samples or objects
}
\description{
This function extract the fields related to the objects or the samples
}
