% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankto_data_utils.R
\name{import.data}
\alias{import.data}
\title{Import data from tsv file}
\usage{
import.data(x, unwanted = NA, taxo_column = "object_annotation_hierarchy")
}
\arguments{
\item{unwanted}{a vector or list of strings containing the names of the categories
you wish to discard. The function will search for these keywords in the field
object_annotation_hierarchy, therefore you can target large categories that include
several species. For example if category A contains categories a,b and c, and
you want to omit a,b and c you can simply put A as argument. By default, the
function will keep all categories.}

\item{file}{one or several .tsv file(s) from a Planktoscope aquisition classified
with Ecotaxa. If multiple files are inserted, make sure they are grouped in a list.}

\item{volumes}{If true, the functions ellipsoid.vol and biovolume will be
called to add two additional columns to the returned data table object.info
containing the volume, calculated based on the ellipsoid approximation and
biovolume of each object}
}
\value{
the result is a list of four data tables. $for.veg to be used by the
vegan package, sample in rows and categories in columns, the entries being
the count of objects being in the corresponding sample and category. $object.info
contains all objects (rows) and their specifications (eg size, color...). $sample.info
contains all samples (rows) and their specifications (eg speed, volume filtered,
concentration factor). $counts is the summary of the absolute and relative
abundances of categories in each sample.
}
\description{
This function is mandatory to use the functions provided in this package.
It will allow you to process the original tsv file and copy the data in
different 'data.table' to be then analysed by specific functions. Within this
function, the functions from biovolume and relative abundance are used to add
these parameters in new columns to the tables
}
