% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbss.R
\name{NBSS.plot}
\alias{NBSS.plot}
\title{Plotting of the NBSS - Normalized Biovolume Size Spectra - of a given sample}
\usage{
NBSS.plot(planktotable, ESD = TRUE)
}
\arguments{
\item{ESD}{If ESD is true, by default, the x axis is expressed as ESD (Eqivalent spherical
diameter for a given volume) in \link{µm}. Otherwise, volumes will be on the x-axis in \link{mm^3}}

\item{objects}{A data table containing objects as rows and details on these objects in the
columns. This argument will be passed in the NBSS function}

\item{samples}{A data table containing samples in rows and details about
the latter in columns.}

\item{sample_name}{A string containing the name of the sample from which to extract objects
and do an NBSS analysis.}
}
\value{
an object of the type ggplot which can then be plotted.
}
\description{
Plotting of the NBSS - Normalized Biovolume Size Spectra - of a given sample
}
\examples{
p <- NBSS.plot(objects=objects,samples=samples,sample_name=sample_id,ESD=TRUE)
p 

}
