% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbss.R
\name{NBSS}
\alias{NBSS}
\title{NBSS : Normalized Biovolume Size Spectra calculation for a given sample,
calculate several NBSS spectra to compare samples}
\usage{
NBSS(objects, samples, sample = samples[1, 1], N = TRUE)
}
\arguments{
\item{objects}{A data table containing objects in rows (not necessarily
from the same sample) and details about the objects in the columns.}

\item{samples}{A data table containing samples in rows and details about
the latter in columns.}

\item{sample}{a string containing the name of the sample for which to calculate the NBSS}
}
\value{
a data table containing two columns, the first with the size spectra,
given in mm^3, the second is the NBSS for each interval of size.
}
\description{
NBSS : Normalized Biovolume Size Spectra calculation for a given sample,
calculate several NBSS spectra to compare samples
}
\note{
The variables used are called biovolumes but to simplify we actually use
volumes. Since biovolumes are the volumes normalized by the same total volume
for all objects in a sample, the result is the same, except that size spectra
are given in mm^3 instead of mm^3/m^3
}
