% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankto_data_utils.R
\name{plot_categ}
\alias{plot_categ}
\title{Taxa selection for plots}
\usage{
plot_categ(x, categ, quantity, ngroups = 9)
}
\arguments{
\item{x}{a long format table obtained using the function get.counts.and.vol()}

\item{categ}{column in the input table containing the category information}

\item{quantity}{column in the input table to quantifuy each category}

\item{ngroups}{maximum number of categories to keep}
}
\value{
Add an extra column to the input table. This column contains the most abundant
categories, the number of categories being defined by ngroups. The other categories are
names "others"
}
\description{
This function add an extra column containing the most abundant categories
}
