% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_abundance.R
\name{relative.abundance}
\alias{relative.abundance}
\title{The relative abundance is calculated simply by dividing the absolute count
of organisms from a certain category and sample from the total number of organisms
in this sample
A column containing the composition is also added, using the formula
composition = sqrt(relative_abundance)}
\usage{
relative.abundance(tab_counts)
}
\arguments{
\item{counts}{a data table with three columns, the sample, the category and
the number of occurences of this category within the sample.}
}
\description{
The relative abundance is calculated simply by dividing the absolute count
of organisms from a certain category and sample from the total number of organisms
in this sample
A column containing the composition is also added, using the formula
composition = sqrt(relative_abundance)
}
\note{
this function does not return anything as it modifies directly the data
table passed as argument.
}
