% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovolumes.R
\name{summed.biovol}
\alias{summed.biovol}
\title{This function computes the sum of biovolumes per category within each sample}
\usage{
summed.biovol(objects)
}
\arguments{
\item{objects}{A data table containing objects in rows (not necessarily
from the same sample) and details about the objects in the columns.}
}
\value{
the result is a data table with : in the rows the different combinations
of samples and categories, and containing an additional column: the summed
biovolumes of each category within their sample.
}
\description{
This function computes the sum of biovolumes per category within each sample
}
