% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_table.R
\name{import.table}
\alias{import.table}
\title{This function is mandatory to use the functions provided in this package.
It will allow you to process the original tsv file and copy the data in
different 'data.table' to be then analysed by specific functions. Within this
function, the functions from biovolume and relative abundance are used to add
these informations to the tables}
\usage{
import.table(file, volumes = TRUE, unwanted = c("not-living"))
}
\arguments{
\item{file}{a .tsv file from a Planktoscope aquisition}

\item{volumes}{If true, the functions ellipsoid.vol and biovolume will be
called to add two additional columns to the returned data table object.info
containing the volume and biovolume of each object}
}
\value{
the result is a list of four data tables. $for.veg to be used by the
vegan package, sample in rows and categories in columns, the entries being
the count of objects being in the corresponding sample and category. $object.info
contains all objects and their specifications (eg size, color...). $sample.info
contains all samples and their specifications (eg speed, volume filtered,
concentration factor). $counts is the summary of the absolute and relative
abundancies of categories in each sample.
}
\description{
This function is mandatory to use the functions provided in this package.
It will allow you to process the original tsv file and copy the data in
different 'data.table' to be then analysed by specific functions. Within this
function, the functions from biovolume and relative abundance are used to add
these informations to the tables
}
